bug - #4 mixed int float arithmetic failed to compile (#16)
Fix issue #4 where mixed `int`/`float` arithmetic produced invalid Rust code, and hardens Incan’s numeric semantics end-to-end (typechecker → lowering → conversions → emitter → stdlib). Numeric behavior is now intentionally **Python-like**, with a single central policy and stdlib-backed runtime helpers.

+++ src/frontend/typechecker/tests.rs
@@ -132,8 +132,9 @@ def foo() -> int:
 
 #[test]
 fn test_arithmetic_division() {
+    // Division always returns float (Python-like semantics)
     let source = r#"
-def foo() -> int:
+def foo() -> float:
   return 10 / 2
 "#;
     assert!(check_str(source).is_ok());
